refactor(types); extract public interfaces for VCalendar
